move signing tools to main package
