+++ src/lib/Zone.spec.ts
@@ -1,10 +1,10 @@
 import { addSeconds, subSeconds } from 'date-fns';
 
-import { ZoneSigner } from '../testUtils/dnssec/ZoneSigner.js';
 import { QUESTION, RECORD, RECORD_TLD } from '../testUtils/dnsStubs.js';
-import type { DnskeyResponse, DsResponse } from '../testUtils/dnssec/responses.js';
-import type { SignatureOptions } from '../testUtils/dnssec/SignatureOptions.js';
 
+import { ZoneSigner } from './utils/dnssec/ZoneSigner.js';
+import type { DnskeyResponse, DsResponse } from './utils/dnssec/responses.js';
+import type { SignatureOptions } from './utils/dnssec/SignatureOptions.js';
 import { DnssecAlgorithm } from './DnssecAlgorithm.js';
 import { Zone } from './Zone.js';
 import { Message } from './dns/Message.js';
